+++ lib/impl/camel_caser/any.ex
@@ -1,0 +1,5 @@
+defimpl Digger.CamelCaser.Protocol, for: Any do
+  def camelize(any, _first_letter) do
+    any
+  end
+end