+++ lib/impl/dasher/any.ex
@@ -1,0 +1,5 @@
+defimpl Digger.Dasher.Protocol, for: Any do
+  def dasherize(any) do
+    any
+  end
+end