+++ lib/impl/lower_caser/atom.ex
@@ -1,0 +1,10 @@
+defimpl Digger.LowerCaser.Protocol, for: Atom do
+  alias Digger.LowerCaser.Delegator
+
+  def lowercase_first(atom) do
+    atom
+    |> to_string
+    |> Delegator.lowercase_first
+    |> String.to_atom
+  end
+end