+++ lib/impl/upper_caser/any.ex
@@ -1,0 +1,5 @@
+defimpl Digger.UpperCaser.Protocol, for: Any do
+  def upcase_first(any) do
+    any
+  end
+end