+++ lib/types.ex
@@ -1,0 +1,9 @@
+defmodule Digger.Types do
+  @moduledoc """
+  A module to document the types used in Digger
+  """
+
+  @type data_type :: any()
+  @type symbol :: atom()
+  @type valid_return_type :: any()
+end